changed interactive mode
