--- login_github
@@ -1,4 +1,0 @@
-#!/bin/bash
-ssh-add ../../githubkey
-ssh -T git@github.com
-bash -i
