+++ simple_shell.c
@@ -1,74 +1,31 @@
 #include "main.h"
-
-void display_prompt()
-{
-	printf("simple_shell_NJR($) ");
-	fflush(stdout);
-}
-
-void execute_command(char *command)
-{
-	pid_t pid = fork();
-
-	if (pid < 0)
-	{
-		perror("Fork error");
-		exit(EXIT_FAILURE);
-	}
-	else if (pid == 0)
-	{
-		char *args[2];
-		args[0] = command;
-		args[1] = NULL;
-
-		execvp(command, args);
-
-		perror("Command execution error");
-		_exit(EXIT_FAILURE);
-	}
-	else
-	{
-		wait(NULL);
-	}
-}
-
-int main()
-{
-	char input[MAX_COMMAND_LENGTH];
-
-	while (1)
-	{
-		display_prompt();
-
-		if (fgets(input, sizeof(input), stdin) == NULL)
-		{
-			printf("\n");
-			break;
-		}
-		input[strcspn(input, "\n")] = 0;
-		
-		execute_command(input);
-	}
-
-	return (0);
-}
+#include <sys/stat.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <unistd.h>
+#include <sys/wait.h>
+#include <sys/types.h>
 
 
 /* Check if a command exists at the given path */
-/**int command_exists(char *cmd)
+int command_exists(char *cmd)
 {
 	struct stat st;
 	return ((stat(cmd, &st) == 0));
-}*/
+}
 
-/* Find a command in the directories specified by the PATH environment variable */
-/**int find_command_in_path(char *cmd, char *fullpath)
+/* 
+ * find_command_in_path - find command in current path
+ * Description: find the path to 
+ * Find a command in the directories specified by the PATH environment variable */
+int find_command_in_path(char *cmd, char *fullpath)
 {
     struct stat st;
     char *path, *token, pth[MAX_PATH_LENGTH];
-*/
+
     /* Check if cmd is an absolute path */
-   /** if (cmd[0] == '/')
+    if (cmd[0] == '/')
     {
         if (stat(cmd, &st) == 0)
         {
@@ -90,17 +47,17 @@ int main()
         sprintf(fullpath, "%s/%s", token, cmd);
         if (stat(fullpath, &st) == 0)
         {
-	  return (1);*/ /* Command found */
-        /**}
+	  return (1); /* Command found */
+        }
         token = strtok(NULL, ":");
     }
 
-    return (0);*/ /* Command not found */
-/*}*/
+    return (0); /* Command not found */
+}
 
 
 /* Handle the 'cat' command */
-/**void handle_cat(char *filename)
+void handle_cat(char *filename)
 {
 	char line[1024];
 	FILE *file = fopen(filename, "r");
@@ -126,13 +83,13 @@ int main()
 	
 	while (1)
 	{
-		*//* Display the prompt only in interactive mode */
-		/**if (is_interactive)
+		/* Display the prompt only in interactive mode */
+		if (is_interactive)
 		{
-			printf("simple_shell_NJR($) ");
-		}*/
+			printf(ANSI_COLOR_RED ANSI_COLOR_RESET);
+		}
 		/* Read a line of input using getline */
-		/**nread = getline(&command, &len, stdin);
+		nread = getline(&command, &len, stdin);
 		if (nread == -1)
 		{
 			free(command);
@@ -140,67 +97,67 @@ int main()
 			{
 				printf("\n");
 			}
-			break;*/ /* Exit on EOF or read error */
-		/*}*/
+			break; /* Exit on EOF or read error */
+		}
 		/* Remove newline character from the end of the command */
-		/**if (command[nread - 1] == '\n')
+		if (command[nread - 1] == '\n')
 		{
 			command[nread - 1] = '\0';
-		}*/
+		}
 		/* Continue if command is empty */
-		/**if (strlen(command) == 0)
+		if (strlen(command) == 0)
 		{
 			continue;
-		}*/
+		}
 		/* Exit the shell if the command is 'exit' */
-		/**if (strcmp(command, "exit") == 0)
+		if (strcmp(command, "exit") == 0)
 		{
 			free(command);
 			exit(0);
-		}*/
+		}
 		/* Split the command into arguments */
-		/**argc = 0;
+		argc = 0;
 		argv[argc] = strtok(command, " ");
 		while (argv[argc] != NULL && argc < MAX_ARGS - 1)
 		{
 			argv[++argc] = strtok(NULL, " ");
 		}
-		argv[argc] = NULL;*/
+		argv[argc] = NULL;
 		
 		/* Handle 'cat' command */
-		/**if (strcmp(argv[0], "cat") == 0 && argv[1] != NULL)
+		if (strcmp(argv[0], "cat") == 0 && argv[1] != NULL)
 		{
 			handle_cat(argv[1]);
 			continue;
-		}*/
+		}
 		/* Check if the command exists in PATH */
-		/**if (!find_command_in_path(argv[0], fullpath))
+		if (!find_command_in_path(argv[0], fullpath))
 		{
 			printf("%s: command not found\n", argv[0]);
 			continue;
-		}*/
+		}
 		
 		/* Fork a new process to execute the command */
-		/**pid = fork();
+		pid = fork();
 		if (pid == -1)
 		{
 			perror("fork");
 			continue;
 		}
 		if (pid == 0)
-		{*/
+		{
 			/* Child process */
-			/**execv(fullpath, argv);*/ /* Execute the command */
-			/*perror("execv");*/ /* Executed only if execv fails */
-			/**exit(EXIT_FAILURE);
+			execv(fullpath, argv); /* Execute the command */
+			perror("execv"); /* Executed only if execv fails */
+			exit(EXIT_FAILURE);
 		}
 		else
-		{*/
+		{
 			/* Parent process */
-			/*wait(NULL);*/ /* Wait for the child process to finish */
-		/**}
+			wait(NULL); /* Wait for the child process to finish */
+		}
 	}
 	free(command);
 	command = NULL;
 	return (0);
-}*/
+}
